% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydraulics.R
\name{desorptivity}
\alias{desorptivity}
\title{Desorptivity calculation}
\usage{
desorptivity(Ks, n, theta_s, theta_r, psi_b, Theta)
}
\arguments{
\item{Ks}{saturated hydraulic conductivity [m/d]}

\item{n}{van Genuchten n parameter [-]}

\item{theta_s}{saturated water content [m3/m3]}

\item{theta_r}{residual water content [m3/m3]}

\item{psi_b}{bubbling pressure [m]}

\item{Theta}{relative saturation [-]}
}
\value{
desorptivity S [m/sqrt(d)]
}
\description{
Calculates desorptivity for stage-1 evaporation
}
