% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydraulics.R
\name{feddes_stress}
\alias{feddes_stress}
\title{Feddes stress function}
\usage{
feddes_stress(theta, theta_star, theta_wp)
}
\arguments{
\item{theta}{current water content [m3/m3]}

\item{theta_star}{optimal water content (no stress) [m3/m3]}

\item{theta_wp}{wilting point [m3/m3]}
}
\value{
stress factor alpha (0-1) [-]
}
\description{
Calculates water stress factor for transpiration
}
\details{
1 above theta_star, 0 below theta_wp, linear in between
}
