% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilmoisture.models-package.R
\docType{package}
\name{sm.drydowns.models-package}
\alias{sm.drydowns.models}
\alias{sm.drydowns.models-package}
\title{soilmoisture.models: DES and muSEC Models for Soil Moisture Drydowns}
\description{
Reference implementation in R of two simplified hydrological models for soil moisture drydowns:
the desorptivity-based model (DES) and the multilayer Surface Evaporative Capacitance extension (muSEC).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dx.doi.org/10.2139/ssrn.5648370}
}

}
\author{
\strong{Maintainer}: Tommaso Martini \email{tommaso.martini@unito.it}

Authors:
\itemize{
  \item Aurora Olivero \email{aurora.olivero@unito.it}
}

}
\keyword{internal}
