% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theta_medio_profondita}
\alias{theta_medio_profondita}
\title{Depth-averaged theta up to z_max}
\usage{
theta_medio_profondita(theta, depths, dz, z_max)
}
\arguments{
\item{theta}{Layer thetas (m3/m3).}

\item{depths}{Layer centers (m).}

\item{dz}{Layer thicknesses (m).}

\item{z_max}{Upper depth (m) for averaging (e.g., 0.30).}
}
\value{
Scalar depth-averaged theta in [0, z_max].
}
\description{
Depth-averaged theta up to z_max
}
