% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydraulics.R
\name{update_theta_analytical}
\alias{update_theta_analytical}
\title{Analytical drainage update}
\usage{
update_theta_analytical(theta0, dt, Ks, dz, theta_r, theta_s, n)
}
\arguments{
\item{theta0}{initial water content [m3/m3]}

\item{dt}{time step [d]}

\item{Ks}{saturated hydraulic conductivity [m/d]}

\item{dz}{layer thickness [m]}

\item{theta_r}{residual water content [m3/m3]}

\item{theta_s}{saturated water content [m3/m3]}

\item{n}{van Genuchten n parameter [-]}
}
\value{
updated water content [m3/m3]
}
\description{
Updates theta using analytical solution for drainage above field capacity
}
